refactor: migrate from core.rag.entities.<submodule> import to from core.rag.entities import
refactor(api): extract OAuthSchema and PluginCredentialType into core/

Move OAuthSchema to core/plugin/entities/oauth.py (keeping list type)
and PluginCredentialType to core/entities/plugin_credential_type.py,
fixing the dependency violation where core/ was importing from services/.

fix

fix

fix
