+++ api/core/plugin/entities/__init__.py
@@ -1,0 +1,5 @@
+from core.plugin.entities.oauth import OAuthSchema
+
+__all__ = [
+    "OAuthSchema",
+]
